+++ alcu_bonus/src/main.c
@@ -6,7 +6,7 @@
 /*   By: lcalvie <lcalvie@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/11 21:32:37 by agautier          #+#    #+#             */
-/*   Updated: 2022/02/13 17:14:40 by lcalvie          ###   ########.fr       */
+/*   Updated: 2022/02/13 17:45:02 by lcalvie          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 static void	play(t_graph *graph)
 {
-	int	x_baton;
-	int	y_baton;
+	double	x_baton;
+	double	y_baton;
 
-	graph->max = list_max(graph->board);
+	graph->max = (double) list_max(graph->board);
 	x_baton = (WIDTH * 0.9) / (graph->max * 2 - 1);
-	y_baton = (HEIGTH * 0.9) / ft_lstsize(graph->board);
+	y_baton = (HEIGTH * 0.9) / ((double)ft_lstsize(graph->board));
 	board_print(graph->board, graph, x_baton, y_baton);
 	usleep(AI_WAIT_TIME);
 	while (graph->board)
